Create all working functions on the calculator
Need to add a way to press equals multiple times and leave the second
operand the same while updating the first operand to be the new total.
Disabled feature for now as it is currently backwards in the case of
multiple 'equals' presses in a row - keeps the first operand the same
and the second operand gets updated from the total.